mitosheet: more error handling and testing
