+++ mitosheet/mitosheet/step_performers/filter.py
@@ -41,19 +41,6 @@ FILTER_CONDITIONS_THAT_REQUIRE_FULL_DATAFRAME = [
     FC_NUMBER_HIGHEST,
 ]
 
-def is_valid_filter_condition(filter_condition: str) -> bool:
-    return filter_condition in [
-        FC_BOOLEAN_IS_FALSE, FC_BOOLEAN_IS_TRUE, FC_DATETIME_EXACTLY,
-        FC_DATETIME_GREATER, FC_DATETIME_GREATER_THAN_OR_EQUAL, FC_DATETIME_LESS,
-        FC_DATETIME_LESS_THAN_OR_EQUAL, FC_DATETIME_NOT_EXACTLY, FC_EMPTY,
-        FC_LEAST_FREQUENT, FC_MOST_FREQUENT, FC_NOT_EMPTY, FC_NUMBER_EXACTLY,
-        FC_NUMBER_GREATER, FC_NUMBER_GREATER_THAN_OR_EQUAL, FC_NUMBER_HIGHEST,
-        FC_NUMBER_LESS, FC_NUMBER_LESS_THAN_OR_EQUAL, FC_NUMBER_LOWEST,
-        FC_NUMBER_NOT_EXACTLY, FC_STRING_CONTAINS, FC_STRING_DOES_NOT_CONTAIN,
-        FC_STRING_ENDS_WITH, FC_STRING_EXACTLY, FC_STRING_NOT_EXACTLY,
-        FC_STRING_STARTS_WITH, FC_STRING_CONTAINS_CASE_INSENSITIVE
-    ]
-
 
 class FilterStepPerformer(StepPerformer):
     """
